requirements.txt
bitstring==3.1.7
SQLAlchemy==1.3.20
python-dotenv~=0.14.0